Change src and dest to be read from bash args
**Me:** Let's change the source and destination to be read from bash args

**ChatGPT:** Sure, here's a modified version of the script that reads the source and destination directories from command-line arguments:

In this version, the script expects the source and destination directories to be provided as command-line arguments when running the script. For example, you could run the script like this:

`./backup.sh /source/ /destination/`

The script will then use the `$1` and `$2` variables to get the values of the first and second command-line arguments, which will be the source and destination directories, respectively.