Surround args with quotes & check presence
**Me:** Don't forget to surround the arguments that are read with quotes to prevent issues. also, add a check to make sure both args are present

**ChatGPT:** Sure, here's a modified version of the script that surrounds the arguments with quotes to prevent issues and adds a check to make sure that both arguments are present:

In this version, the `source` and `destination` variables are surrounded with double quotes to prevent any issues with whitespace in the arguments. The script also checks if the `source` and `destination` variables are empty, and if so, prints an error message and exits with an error code. This ensures that the script will only run if both arguments are provided.