Only check the dest drive
**Me:** Can we change the check for the drive being full to only check the drive of the destination path that the user gave

**ChatGPT:** Sure, here's a modified version of the script that only checks the drive of the destination path that the user provided:

In this version, the script uses the `df` command to get the drive of the destination directory, and then checks if that drive is 100% full. If it is, the script prints a snarky message and exits with an error code. This ensures that the script only checks the drive of the destination directory provided by the user, rather than checking all the drives on the system.